refs #7 Stereotype Recommendation Algorithm
Modified the item in order to include the proximity to the current determined stereotype, which is used by the algorithm to sort the map in descending order to get the most relevant items.

Modified the Stereotypes, such that they look a lit bit better and all entries in the attributeProbabilityMap only take one line and not two, which reduced the readability.

Also fixed a bug in the database query, which prevented Unisex items from being read from the database.